requirements.txt
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
coverage==7.5.4
cryptography==42.0.8
daphne==4.1.2
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.0.6
django-debug-toolbar==4.4.2
django-extensions==3.2.3
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
exceptiongroup==1.2.1
executing==2.0.1
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.3.1
iniconfig==2.0.0
ipython==8.25.0
jedi==0.19.1
lazy-object-proxy==1.10.0
matplotlib-inline==0.1.7
msgpack==1.0.8
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pluggy==1.5.0
prompt_toolkit==3.0.46
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
PySocks==1.7.1
pytest==7.4.4
pytest-assert-utils==0.3.1
pytest-common-subject==1.1.1
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-drf==1.1.3
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
python3-openid==3.2.0
pytils==0.4.1
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.22.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.0
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
urllib3==2.2.2
uvicorn==0.30.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
zope.interface==6.4.post2
